requirements.txt
spacy==3.6.0
scikit-learn==1.5.0
numpy==1.25.0
python-dotenv
Flask==3.1.2  # For web interface